refactor(themes): rename atom-one-light to one-light
